+++ literal.h
@@ -1,14 +1,14 @@
-class Literal: public Operand
+class Literal : public Operand
 {
 public:
-    Literal(int value)
-    {
-        this->value = value;
-    }
-    int evaluate()
-    {
-        return value;
-    }
+	Literal(double value)
+	{
+		this->value = value;
+	}
+	double evaluate()
+	{
+		return value;
+	}
 private:
-    int value;
+	double value;
 };
