+++ minus.h
@@ -1,11 +1,11 @@
-class Minus: public SubExpression
+class Minus : public SubExpression
 {
 public:
-    Minus(Expression* left, Expression* right): SubExpression(left, right)
-    {
-    }
-    int evaluate()
-    {
-        return left->evaluate() - right->evaluate();
-    }
+	Minus(Expression* left, Expression* right) : SubExpression(left, right)
+	{
+	}
+	double evaluate()
+	{
+		return left->evaluate() - right->evaluate();
+	}
 };
