+++ operand.cpp
@@ -1,5 +1,6 @@
 #include <cctype>
 #include <iostream>
+#include <sstream>
 #include <list>
 #include <string>
 
@@ -14,22 +15,22 @@ using namespace std;
 
 Expression* Operand::parse(stringstream &in)
 {
-    char paren;
-    int value;
+	char paren;
+	double value;
 
-    in >> ws;
-    if (isdigit(in.peek()))
-    {
-        in >> value;
-        Expression* literal = new Literal(value);
-        return literal;
-    }
-    if (in.peek() == '(')
-    {
-        in >> paren;
-        return SubExpression::parse(in);
-    }
-    else
-        return new Variable(parseName(in));
-    return 0;
+	in >> ws;
+	if (isdigit(in.peek()))
+	{
+		in >> value;
+		Expression* literal = new Literal(value);
+		return literal;
+	}
+	if (in.peek() == '(')
+	{
+		in >> paren;
+		return SubExpression::parse(in);
+	}
+	else
+		return new Variable(parseName(in));
+	return 0;
 }
